Change statemachine.go to host.go and move tests to host_test.go
